Build organelle screen now fully functional
Opening the build organelle panel through the cell panel now displays a list of additional organelles that you can buy, which are then enabled in the cell if you choose to build them. These items are greyed out with a message on top of it if the organelle is locked due to the cell's level not being high enough, if the organelle has already been purchased, and (to be added when cell membranes are added) if the cell membrane isn't high enough level.