Last minute solution for missing save files on first execution
When the user runs the executable for the first time, there are no save files. I quickly threw together a bandaid solution that tells the user to manually put the default save files in the right spot on their computer before continuing. Preferably it would generate those files on its own but alas, we're out of time.